Add issues to a project based on milestone
The action may now be triggered when an issue is either labeled or added
to a milestone. Use the 'milestone-name' input to define the milestone
title to match. label-name and milestone-name are mutually exclusive.

Example:

This YAML is meant to be triggered whenever an issue has been milestoned.

```
on:
  issues:
    types: [milestoned]
jobs:
  Move_Milestoned_Issue_On_Project_Board:
    runs-on: ubuntu-latest
    steps:
    - uses: konradpabjan/actions-move-labeled-issue-repo@v1.0
      with:
        action-token: "${{ secrets.MY_TOKEN }}"
        project-url: "https://github.com/orgs/github/projects/1"
        column-name: "On Deck"
        milestone-name: "v1.0"
        columns-to-ignore: "In Review,Ready to deploy,Done"
 ```